+++ src/kadrx/KaDrxPub.h
@@ -87,7 +87,7 @@ class KaDrxPub : public QThread {
         /// @param buf The raw buffer of data from the downconverter
         /// channel. It contains all Is and Qs
 		/// @param pulsenum The pulse number. Will be zero for raw data.
-        void _publishDDS(char* buf, unsigned int pulsenum);
+        void _publishDDS(char* buf, long long pulsenum);
         
         /**
          * Return true iff the current configuration is valid.
